backend/requirements.txt
# Core Django
Django==4.2.7
django-ninja==1.0.1
django-cors-headers==4.3.1

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1

# Data Processing & ML
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# System & Performance
psutil==5.9.6

# Reporting
reportlab==4.0.7

# Configuration
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# External APIs
plaid-python==16.0.0

# Production Server
gunicorn==21.2.0

# Security & Rate Limiting
django-ratelimit==4.1.0
django-environ==0.11.2

# Authentication & JWT
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.2.1
google-auth==2.23.4
httpx==0.25.1
email-validator==2.3.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2